refactor: 전역 사용 커스텀 어노테이션 패키지 경로 변경
