+++ backend/src/test/java/woowacourse/touroot/travelogue/controller/TravelogueControllerTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import woowacourse.touroot.global.AcceptanceTest;
 import woowacourse.touroot.utils.DatabaseCleaner;
 import woowacourse.touroot.utils.TestFixture;
